Publishing Arjana Project on Github
Arjana is a sanskrit word which means achievement. This game is based on three concepts - Perceive, Believe and Achieve. It is a small act of a real game called Life.

It is a one of a kind Motivational game that contains hidden meanings, a set of quotes and poems based on situations faced by Students, Employees, Entreprenuers and ****. This game is full of symbolism, its on the player to find the meaning and apply it in their Life.

Ultimately in Arjana, you have to find out the identity of the last person. The one that resides in all of us. Its not that easy and not what you are thinking of. Initially only Student level is for play.

Player has to win the current level in order to unlock the next level. In levels, you have to collect all the gems & crack the puzzle at last.The puzzle is based on gems that contains quotes, out of which only one is the key to unlock the next level.

Gameplay may look easy, but your patience & thinking process will be challenged. If you the One, then you are the Ultimate.